requirements.txt
python-binance==1.0.19
python-dotenv==1.0.0
requests==2.31.0
numpy==1.26.2
fastapi==0.115.0
uvicorn==0.30.0
prisma==0.15.0
pytest==8.0.0
pytest-asyncio==0.23.5
pytest-cov==4.1.0
httpx==0.27.0
